Fixing stuff

+++ src/app/models/file.js
@@ -1,32 +1,13 @@
 const { date } = require('../lib/utils')
 const db = require('../config/db')
 const fs = require('fs')
+const Base = require('./Base')
+
+Base.init({ table: 'files' })
 
 
 module.exports = {
-    create({ filename, path }) {
-        const query = `
-            INSERT INTO files (
-                name,
-                path
-                
-                
-            )
-            VALUES ($1, $2)
-            RETURNING id
-            
-          
-        `
-
-        const values = [
-            filename,
-            path
-
-        ]
-
-
-        return db.query(query, values)
-    },
+    ...Base,
     join({ recipe_id, file_id }) {
         const query = `
         INSERT INTO recipe_files(
@@ -60,9 +41,18 @@ module.exports = {
 
             const file = result.rows[0]
             
-            fs.unlinkSync(file.path)
 
-            
+            if (file != undefined && fs.existsSync(file.path) == true) {
+
+                fs.unlinkSync(file.path)
+
+            }
+
+
+            return await db.query(`DELETE FROM recipe_files
+        WHERE recipe_files.id = $1
+        RETURNING file_id`, [id])
+
         } catch (err) {
             console.error(err)
         }
@@ -73,11 +63,8 @@ module.exports = {
 
 
 
-        return db.query(`DELETE FROM recipe_files
-        WHERE recipe_files.id = $1
-        RETURNING file_id`, [id])
     },
-    
+
 
 
 }
